frontend/requirements.txt
nltk
networkx
streamlit
PyPDF2
scipy
python-docx
pandas
matplotlib
wordcloud